move pg metrics out to metrics.go
